Added remove and backup functionality for archives

+++ Module1.vb
@@ -4,7 +4,7 @@ Module Module1
     ' Declared variables for iHistorian_SDK
     Public ConnectedServer As iHistorian_SDK.Server
     'Input Variables
-    Public ParameterFlags(8, 1) As String
+    Public ParameterFlags(11, 1) As String
     Public ServerName As String = System.Net.Dns.GetHostName
     Public FilePath As String
     Public Username As String = Nothing
@@ -13,14 +13,27 @@ Module Module1
     Public Overwrite As Boolean = False
     'Other Variables
     Public ArchivePath As String = Nothing
+    Public BackupPath As String = Nothing
     Public ProgramFilesDirs(1) As String
     Public ArchiveBackupApp As String = "C:\Program Files\Proficy\Proficy Historian\x86\Server\ihArchiveBackup.exe"
+    Public RestoreAction As Boolean = False
+    Public BackupAction As Boolean = False
+    Public RemoveAction As Boolean = False
+    Public ExistingArchiveFiles As Object
+
+    Dim mydatastores As iHistorian_SDK.DataStores
+    Dim mydatastore As iHistorian_SDK.DataStore
+    Dim myarchives As iHistorian_SDK.Archives
+    Dim myarchive As iHistorian_SDK.Archive = Nothing
 
     Sub Main(args As String())
         ParameterFlags = {
             {"-S", "The hostname of the server for connecting to"},
             {"-U", "The username for authenticating with historian (leave blank to use AD group)"},
             {"-P", "The password for authenticating with historian (leave blank to use AD group)"},
+            {"--RESTORE", "Restoring archives listed in the file"},
+            {"--BACKUP", "Backing up archives listed in the file"},
+            {"--REMOVE", "Removing archives listed in the file"},
             {"-F", "The file path to the file containing the list of the archive paths to import for restoring (only IHAs at the moment), one item per line"},
             {"-O", "Overwrite any existing IHA files in the default archive path"},
             {"-D", "The specified datastore (leave blank to use the default datastore)"},
@@ -36,42 +49,49 @@ Module Module1
             GoTo errc
         End If
         For i As Integer = 0 To args.Count() - 1
+
             Select Case UCase(args(i))
                 Case ParameterFlags(0, 0) '-s
                     If (i < args.Count() - 1) Then
-                        If Not IsInArray(CStr(args(i + 1)), ParameterFlags) Then
+                        If (IsInArray(CStr(args(i + 1)), ParameterFlags, True) = -1) Then
                             ServerName = args(i + 1)
                         End If
                     End If
                 Case ParameterFlags(1, 0) '-u
                     If (i < args.Count() - 1) Then
-                        If Not IsInArray(CStr(args(i + 1)), ParameterFlags) Then
+                        If (IsInArray(CStr(args(i + 1)), ParameterFlags, True) = -1) Then
                             Username = args(i + 1)
                         End If
                     End If
                 Case ParameterFlags(2, 0) '-p
                     If (i < args.Count() - 1) Then
-                        If Not IsInArray(CStr(args(i + 1)), ParameterFlags) Then
+                        If (IsInArray(CStr(args(i + 1)), ParameterFlags, True) = -1) Then
                             Password = args(i + 1)
                         End If
                     End If
-                Case ParameterFlags(3, 0) '-f
+                Case ParameterFlags(3, 0) '--restore
+                    RestoreAction = True
+                Case ParameterFlags(4, 0) '--backup
+                    BackupAction = True
+                Case ParameterFlags(5, 0) '--remove
+                    RemoveAction = True
+                Case ParameterFlags(6, 0) '-f
                     If (i < args.Count() - 1) Then
-                        If Not IsInArray(CStr(args(i + 1)), ParameterFlags) Then
+                        If (IsInArray(CStr(args(i + 1)), ParameterFlags, True) = -1) Then
                             FilePath = args(i + 1)
                         End If
                     End If
-                Case ParameterFlags(4, 0) '-o
+                Case ParameterFlags(7, 0) '-o
                     Overwrite = True
-                Case ParameterFlags(5, 0) '-d
+                Case ParameterFlags(8, 0) '-d
                     If (i < args.Count() - 1) Then
-                        If Not IsInArray(CStr(args(i + 1)), ParameterFlags) Then
+                        If (IsInArray(CStr(args(i + 1)), ParameterFlags, True) = -1) Then
                             Datastore = args(i + 1)
                         End If
                     End If
-                Case ParameterFlags(6, 0) '-nc
+                Case ParameterFlags(9, 0) '-nc
                     ArchiveBackupApp = ""
-                Case ParameterFlags(7, 0), ParameterFlags(8, 0), ParameterFlags(9, 0) '-help
+                Case ParameterFlags(10, 0), ParameterFlags(8, 0), ParameterFlags(9, 0) '-help
                     OutputHelp()
                     GoTo errc
                 Case Else
@@ -80,9 +100,15 @@ Module Module1
         Next
 
         'check for the Historian SDK DLL Depedencies
-        If Not System.IO.File.Exists("C:\Windows\SYSWOW64\iHSDK.dll") Then
+        If (Not System.IO.File.Exists("C:\Windows\SYSWOW64\iHSDK.dll")) And (Not System.IO.File.Exists("C:\Windows\System32\iHSDK.dll")) Then
             Console.WriteLine("The Historian SDK has not been installed, please install this from the Historian Installation Disk under Client Tools")
             Console.WriteLine("The ISO image can be downloaded from the GE Website, https://digitalsupport.ge.com/en_US/Download/Historian-7-1")
+            GoTo errc
+        End If
+
+        'check that if an action is given as a param, that only one param is specified
+        If (BackupAction And RestoreAction) Or (RestoreAction And RemoveAction) Then
+            Console.WriteLine("Cannot have more than one action in a single commmand - except backing up and removing archives at the same time")
             GoTo errc
         End If
 
@@ -104,7 +130,6 @@ Module Module1
                 dirInfo = New IO.DirectoryInfo(ProgramFileDir)
                 SearchForFileInDirectory(matches, dirInfo, "ihArchiveBackup.exe")
                 If matches.Count > 0 Then
-                    Console.WriteLine("found")
                     ArchiveBackupApp = matches(0).FullName
                     Exit For
                 End If
@@ -133,29 +158,39 @@ Module Module1
             GoTo errc
         End If
         lEndTime = Timer
-        Console.WriteLine("Connect completed in " & lEndTime - lStartTime & " seconds")
+        Console.WriteLine("Connection completed in " & lEndTime - lStartTime & " seconds")
         If Not CheckConnection() Then
             Console.WriteLine("Connection to server " + ServerName + " is bad.")
         Else
-            Console.WriteLine("Proficy Historian SDK Sample - Connected to: " + ServerName)
+            Console.WriteLine("HistorianArchiveCLI - Connected to: " + ServerName)
         End If
 
-        'Check datastore archive path
-        Dim mydatastores As iHistorian_SDK.DataStores
-        Dim mydatastore As iHistorian_SDK.DataStore
+        'Check datastore archive path, backup path, and loading archives
         mydatastores = ConnectedServer.DataStores
         For Each mydatastore In mydatastores.Item    ' Iterate through each element.  
             If String.IsNullOrEmpty(Datastore) And mydatastore.IsDefault Then
                 ArchivePath = mydatastore.Archives.ArchivingOptions("ArchiveDefaultPath")
+                BackupPath = mydatastore.Archives.ArchivingOptions("ArchiveBackupPath")
+                myarchives = ConnectedServer.Archives
+                Exit For
             ElseIf mydatastore.Name = Datastore Then    ' If datastore name equals the datastore we're after
                 ArchivePath = mydatastore.Archives.ArchivingOptions("ArchiveDefaultPath")
-                Exit For    ' Exit loop. 
+                BackupPath = mydatastore.Archives.ArchivingOptions("ArchiveBackupPath")
+                myarchives = mydatastore.Archives
+                Exit For
             End If
         Next
 
         'check the archive path
         If String.IsNullOrEmpty(ArchivePath) Then
             Console.WriteLine("The archive path could not be found, please check you've provided the correct datastore name," +
+                              " or that there is only 1 datastore on the server")
+            GoTo errc
+        End If
+
+        'check the backup path
+        If (BackupAction And String.IsNullOrEmpty(BackupPath)) Then
+            Console.WriteLine("The backup path could not be found, please check you've provided the correct datastore name," +
                               " or that there is only 1 datastore on the server")
             GoTo errc
         End If
@@ -170,7 +205,7 @@ Module Module1
         If Not String.IsNullOrEmpty(ArchiveBackupApp) Then
             Console.WriteLine("Backing up IHC file before commencing restoration")
             Dim p As New Process
-            Dim psi As New ProcessStartInfo("C:\Program Files\Proficy\Proficy Historian\x86\Server\ihArchiveBackup.exe", " -s " + ServerName + " -t 30 -c")
+            Dim psi As New ProcessStartInfo(ArchiveBackupApp, " -s " + ServerName + " -t 30 -c")
             psi.CreateNoWindow = False
             psi.UseShellExecute = False
             p.StartInfo = psi
@@ -178,50 +213,26 @@ Module Module1
             p.WaitForExit()
         End If
 
-        'Loop through each line of the file
-        Dim myarchives As iHistorian_SDK.Archives
-        Dim myarchive As iHistorian_SDK.Archive = Nothing
-        myarchives = ConnectedServer.Archives
+        'check if backup or remove action are to be performed, if so get the list of files, and place in variable
+        If (BackupAction Or RemoveAction) Then
+            ReDim ExistingArchiveFiles(myarchives.Count)
+            Dim i As Integer = 0
+            For Each myarchive In myarchives.Item
+                ExistingArchiveFiles(i) = myarchive.FileName.ToUpper()
+                i = i + 1
+            Next
+        End If
+
         For Each Line As String In File.ReadLines(FilePath)
             If Not System.IO.File.Exists(Line) Then
                 Console.WriteLine("The archive " + Line + " doesn't exist")
             Else
-                'check if the archive path is inside the archive path (or if it needs to be copied)
-                If Line.Contains(ArchivePath) Then
-                    If Not IsFileInUse(ArchivePath + Path.GetFileName(Line)) Then
-                        myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(Line), Line, 0, Datastore)
-                    Else
-                        Console.WriteLine("The archive " + Line + " is currently in use, and cannot be overwritten.")
-                    End If
-
-                Else 'copy the file
-
-                    If Not Overwrite Then
-                        'check if a file already exists in the archive path
-                        If Not System.IO.File.Exists(ArchivePath + Path.GetFileName(Line)) Then
-                            System.IO.File.Copy(Line, ArchivePath + Path.GetFileName(Line))
-                            myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(Line), ArchivePath + Path.GetFileName(Line), 0, Datastore)
-                        Else
-                            Console.WriteLine("File " + Line + " already exists in the default archive path, " + ArchivePath)
-                        End If
-                    Else
-                        'check if the file is in use
-                        If Not IsFileInUse(ArchivePath + Path.GetFileName(Line)) Then
-                            System.IO.File.Copy(Line, ArchivePath + Path.GetFileName(Line), True)
-                            myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(Line), ArchivePath + Path.GetFileName(Line), 0, Datastore)
-                        Else
-                            Console.WriteLine("The archive " + ArchivePath + Path.GetFileName(Line) + " is currently in use, and cannot be overwritten.")
-                        End If
-                    End If
-                End If
-
-                If myarchive Is Nothing Then
-                    Console.WriteLine("An Error Occured Trying To Restore The Archive. " + Line +
-                    " The Details Of The Error Follow: " + ConnectedServer.Archives.LastError)
+                If RestoreAction Then
+                    RestoreArchive(Line)
                 Else
-                    Console.WriteLine("The archive " + Line + " has been successfully added")
+                    If BackupAction Then BackupArchive(Line)
+                    If RemoveAction Then RemoveArchive(Line)
                 End If
-                myarchive = Nothing
             End If
         Next
 
@@ -258,25 +269,33 @@ errc:
         CheckConnection = False
     End Function
 
-    Private Function IsInArray(valToBeFound As Object, arr As Object) As Boolean
+    Private Function IsInArray(valToBeFound As Object, arr As Object, TwoDim As Boolean) As Integer
         'DEVELOPER: Ryan Wells (wellsr.com)
         'DESCRIPTION: Function to check if a value is in an array of values
         'INPUT: Pass the function a value to search for and an array of values of any data type.
         'OUTPUT: True if is in array, false otherwise
-        Dim i As Integer
+        Dim i As Integer = -1
         On Error GoTo IsInArrayError 'array is empty
+        IsInArray = -1
 
         For i = LBound(arr, 1) To UBound(arr, 1)
-            If arr(i, 0) = valToBeFound Then
-                IsInArray = True
-                Exit Function
+            If TwoDim Then
+                If arr(i, 0) = valToBeFound Then
+                    IsInArray = i
+                    Exit Function
+                End If
+            Else
+                If arr(i) = valToBeFound Then
+                    IsInArray = i
+                    Exit Function
+                End If
             End If
         Next i
 
         Exit Function
 IsInArrayError:
         On Error GoTo 0
-        IsInArray = False
+        IsInArray = -1
     End Function
 
     Public Function IsFileInUse(sFile As String) As Boolean
@@ -333,5 +352,104 @@ IsInArrayError:
         End Try
     End Sub
 
+    'Restore Archive Function
+    Sub RestoreArchive(fileName As String)
+        'check if the archive path is inside the archive path (or if it needs to be copied)
+        If fileName.Contains(ArchivePath) Then
+            If Not IsFileInUse(ArchivePath + Path.GetFileName(fileName)) Then
+                myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(fileName), fileName, 0, Datastore)
+            Else
+                Console.WriteLine("The archive " + fileName + " is currently in use, and cannot be overwritten.")
+            End If
+
+        Else 'copy the file
+            If Not Overwrite Then
+                'check if a file already exists in the archive path
+                If Not System.IO.File.Exists(ArchivePath + Path.GetFileName(fileName)) Then
+                    System.IO.File.Copy(fileName, ArchivePath + Path.GetFileName(fileName))
+                    myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(fileName), ArchivePath + Path.GetFileName(fileName), 0, Datastore)
+                    'check if file is a zip file
+
+                    If (Path.GetExtension(fileName).ToUpper() = ".ZIP") Then
+                        If System.IO.File.Exists(ArchivePath + Path.GetFileName(fileName)) Then
+                            If Not IsFileInUse(ArchivePath + Path.GetFileName(fileName)) Then
+                                System.IO.File.Delete(ArchivePath + Path.GetFileName(fileName))
+                            End If
+                        End If
+                    End If
+
+                Else
+                    Console.WriteLine("File " + fileName + " already exists in the default archive path, " + ArchivePath)
+                End If
+            Else
+                'check if the file is in use
+                If Not IsFileInUse(ArchivePath + Path.GetFileName(fileName)) Then
+                    System.IO.File.Copy(fileName, ArchivePath + Path.GetFileName(fileName), True)
+                    myarchive = myarchives.Add(Path.GetFileNameWithoutExtension(fileName), ArchivePath + Path.GetFileName(fileName), 0, Datastore)
+                Else
+                    Console.WriteLine("The archive " + ArchivePath + Path.GetFileName(fileName) + " is currently in use, and cannot be overwritten.")
+                End If
+            End If
+        End If
+
+        If myarchive Is Nothing Then
+            Console.WriteLine("An Error Occured Trying To Restore The Archive. " + fileName +
+            " The Details Of The Error Follow: " + ConnectedServer.Archives.LastError)
+        Else
+            Console.WriteLine("The archive " + fileName + " has been successfully added")
+        End If
+        myarchive = Nothing
+    End Sub
+
+    'Backup Archive Function
+    Sub BackupArchive(fileName As String)
+        'Find the filename in the existing archive list
+        Dim idx = IsInArray(CStr(fileName).ToUpper(), ExistingArchiveFiles, False)
+        'if it does exist
+        If (idx <> -1) Then
+            'verify the archive filename directly
+            myarchive = myarchives.Item(idx)
+            'if the filename also matches
+            If (myarchive.FileName.ToUpper() = CStr(fileName).ToUpper()) Then
+                If myarchive.Backup(BackupPath + Path.GetFileNameWithoutExtension(fileName), Datastore) Then
+                    Console.WriteLine("The archive " + fileName + " has been successfully backed up")
+                Else
+                    Console.WriteLine("The archive " + fileName + " was unable to be backed up")
+                End If
+            Else
+                Console.WriteLine("The archive " + fileName + " does not exist on the server, or has encountered an error, try rerunning")
+            End If
+        Else
+            Console.WriteLine("The archive " + fileName + " does not exist on the server")
+        End If
+    End Sub
+
+    'Remove Archive Function
+    Sub RemoveArchive(fileName As String)
+        'Find the filename in the existing archive list
+        Dim idx = IsInArray(CStr(fileName).ToUpper(), ExistingArchiveFiles, False)
+        'if it does exist
+        If (idx <> -1) Then
+            'verify the archive filename directly
+            myarchive = myarchives.Item(idx)
+            'if the filename also matches
+            If (myarchive.FileName.ToUpper() = CStr(fileName).ToUpper()) Then
+                'check if the archive is current
+                If (myarchive.IsCurrent) Then
+                    Console.WriteLine("The archive " + fileName + " is set to current, and cannot be removed")
+                Else
+                    If myarchive.Delete(Datastore) Then
+                        Console.WriteLine("The archive " + fileName + " has been successfully removed")
+                    Else
+                        Console.WriteLine("The archive " + fileName + " was unable to be removed")
+                    End If
+                End If
+            Else
+                Console.WriteLine("The archive " + fileName + " does not exist on the server, or has encountered an error, try rerunning")
+            End If
+        Else
+            Console.WriteLine("The archive " + fileName + " does not exist on the server")
+        End If
+    End Sub
 
 End Module
